aibackend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
Jinja2
python-multipart
psycopg2-binary
pgvector
tqdm
openai
requests
